COMP:powf() is not defined on old Sun. Anyway, computing in double and store the result in float is a better strategy.

+++ VTK/VolumeRendering/vtkOpenGLGPUVolumeRayCastMapper.cxx
@@ -316,7 +316,8 @@ public:
             {
             if(*ptr>0.0001f)
               {
-              *ptr=1.0f-powf(1.0f-(*ptr),static_cast<float>(factor));
+              *ptr=static_cast<float>(1.0-pow(1.0-static_cast<double>(*ptr),
+                                              factor));
               }
             ++ptr;
             ++i;
